+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 3.0.0
+
+- ESLint 10 support.
+- Plugin eslint-plugin-promise has been removed.
+- Updated eslint-plugin-perfectionist to v5.
+- Updated globals to v17.
+- Added a new argument to the testPluginConfig function for testing the use of
+  recommended rules. Type `recommendedRules?: Linter.RulesRecord`.
+- Rules changes:
+  - radix - changed config (eslint 10).
+  - no-useless-assignment - enabled (recommended rule).
+  - no-debugger - enabled (recommended rule).
+  - init-declarations - disabled (conflicts with no-useless-assignment).
+  - perfectionist/sort-import-attributes - enabled (new rule).
+  - perfectionist/sort-export-attributes - enabled (new rule).
+  - @typescript-eslint/strict-void-return - enabled (new rule).
+  - @typescript-eslint/no-useless-default-assignment - enabled (new rule).
+  - @typescript-eslint/class-literal-property-style - changed config (getters
+    are better for libraries).
+
 ## 2.5.0
 
 - Added